requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
asgi-redis==1.1.0
asgiref==1.0.1
aspy.yaml==0.2.2
attrs==16.3.0
autobahn==0.18.0
Automat==0.5.0
bandit==1.4.0
billiard==3.3.0.23
boltons==16.1.1
brotlipy==0.6.0
cached-property==1.3.0
celery==3.1.23
certifi==2017.1.23
cffi==1.9.1
channels==1.1.1
click==6.7
constantly==15.1.0
coverage==4.3.4
daphne==1.1.0
decorator==4.0.11
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.10.6
django-allauth==0.31.0
django-authtools==1.5.0
django-braces==1.11.0
django-crispy-forms==1.6.1
django-log-request-id==1.3.2
django-model-utils==2.6.1
django-naomi==0.8
django-webpack-loader==0.4.1
django.js-vinta==0.8.2.dev2
easy-thumbnails==2.3
fixmydjango==0.2
flake8==2.6.2
flake8-putty==0.4.0
gitdb2==2.0.0
GitPython==2.1.3
gunicorn==19.7.0
incremental==16.10.1
ipython==5.3.0
ipython-genutils==0.2.0
isort==4.2.5
jsonschema==2.6.0
kombu==3.0.37
mccabe==0.5.3
model-mommy==1.3.2
msgpack-python==0.4.8
nodeenv==1.1.2
oauthlib==2.0.1
olefile==0.44
opbeat==3.5.2
packaging==16.8
pbr==2.0.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
pre-commit==0.13.3
prompt-toolkit==1.0.13
psycopg2==2.7.1
ptyprocess==0.5.1
pycodestyle==2.0.0
pycparser==2.17
pyflakes==1.2.3
Pygments==2.2.0
pyparsing==2.2.0
python-decouple==3.0
python3-openid==3.1.0
pytz==2016.10
PyYAML==3.12
redis==2.10.5
requests==2.13.0
requests-oauthlib==0.8.0
safety==0.6.0
simplegeneric==0.8.1
six==1.10.0
smmap2==2.0.1
stevedore==1.21.0
termcolor==1.1.0
traitlets==4.3.2
Twisted==17.1.0
txaio==2.6.1
urllib3==1.20
vine==1.1.3
virtualenv==15.1.0
wcwidth==0.1.7
whitenoise==3.3.0
zope.interface==4.3.3
